backend/requirements.txt
Django
django-jazzmin
whitenoise
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-cors-headers
django-allauth
dj-rest-auth
python-decouple
django-ratelimit
django-axes
psycopg2-binary
requests
django-environ
razorpay
google-auth
xhtml2pdf
gunicorn